Fixes MODsuits Not Being Able To Deploy Individual Parts (#86511)
Replaces the check for the mod_parts list with the get_parts() proc
which converts that list into the parts of the MODsuit. Which is how all
the other checks do it. This was changed in #86031 from just checking
the parts list to checking mod_parts. From my testing, checking
contents, parts, and get_parts() all make it work as intended. I felt
like the proc was the proper way to fix this issue. There might be a
reason it was changed to use mod_parts specifically but I cannot find
it.
I have been playing c*rbon recently and I think that wearing just the
chest piece of a modsuit is drippy AF and I want to be able to do that.
:cl:
fix: You can deploy MODsuit parts individually again.
/:cl:

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>