Adds a Quiver to the Syndie rebar + various rebar  and bolt improvements. (#86116)
Adds a quiver to the traitor "Syndicate Rebar" kit, can be used to store
rods and reload the Rebar more efficiently.

![image](https://github.com/user-attachments/assets/bedc98d5-4b2d-4a68-b598-0853c17d73cb)

Fixes Rebar bows having more ammo than intended.

Rebar bows will now draw their string with left click instead of just
the use button.

Quivers are now a neck storage item.

Hydrogen bolts have lost their ability to embed and pierce an infinite
number of targets in exchange for a small damage buff , increased
accuracy on bodyparts and the ability to pierce walls.

This PR is the fruit of a discussion spanned over the course of several
months between me and KingKumaArt (the creator of the rebar bows.), He's
now busy abroad, so I'm pushing these changes in his stead.

**Changes to the syndie rebar and quivers**

The syndicate rebar bow is a fun albeit slightly underwhelming weapon,
at least compared to its direct competitor, the revolver.

The numbers prove this, as according to the charts it's very rarely
bought on LRP and doesn't even show on the MRP charts.

https://superset.moth.fans/superset/dashboard/4/?native_filters_key=cBYTbRt2At8pt1VkwV8O7NEZZZeDiIQ5fZgSMfjoICB0yow1BWJXDG5SyV9MJKTf

I think most of the issues with the weapon can be attributed to the
horrible jank that surrounds it.

Reloading and storing ammunition for this weapon is a management
nightmare, as the rods cannot be stacked and have to be manually loaded
by hand, making it horrendous to use in combat.

While that's fine for the makeshift crew version, Traitor weapons should
be more realiable in a combat scenario.

**So i'm squashing both problems by introducing a traitor version of the
quiver.**

It comes packed with the box you get the crossbow from, can be worn on
the neck, stacks up to 20 rods, and features an action button that
allows you to slowly chamber your rods into the rebar.

Video demonstration: https://www.youtube.com/watch?v=nG9efFOyauw

The crew version has also been moved to a neck storage item, but has
half the capacity and lacks the ability to quick reload.

I've also made the firing process slightly less annoying by having left
click draw and undraw the string; as it stands, you have to alternate
left click and use on hand after each shot, which is just jarring.

Created 2 new variables to replace the ALWAYS_HIT_TRAIT, so we don't
have to slap a new trait every time we want to dictate how accurate a
gun or projectile should be at range.

Lastly, I fixed an issue that allowed the bow to fit 1 extra bolt than
its intended capacity. if you were to chamber your last rod, it does end
up nerfing the burst of the syndie variant, as it drops down from 4 to
3, but the smoother reload should make up for it.

**Hydrogen bolts**

As it stands, all the atmosian bolt variants occupy different niches and
they are more or less balanced.

Except for the hydrogen bolts.

With zauker bolts doing 60 toxin damage, guaranteed embed, and Syndie
bolts doing 55 brute + the high delimb chance, Hydrogen bolts just don't
have any real purpose to exist with their lower damage and lack of the
ability to delimb.

Metallic hydrogen in general doesn't have any good applications, the
armor is weaker than your run of the mill sec vest, the axe is a
downgrade over the roundstarting one....; I could go on, point being
that something that requires in depth atmos knowledge and 30-40 minutes
of hard work shouldn't be outclassed by what's readily available at the
start of the shift.

So what if we made the hydrogen bolts go through walls?

Long range weaponry is a niche that remains mostly unfilled in the
traitor arsenal.

The syndicate rebar technically does have a scope, but requiring direct
line of sight makes it unreliable and far too risky to use, since you
could be jumped at any times while you are zoomed in.

So i've made the hydrogen bolts a weaker version of the nukie penetrator
rounds, with slightly lower damage, no infinite pierce and inability to
damage the AI core.

I've also improved their long range accuracy, as these bolts are meant
to be accurate at longer ranges, (they drop off massively past the 15
tiles)

Now yes, the bolts will be accessible to the crew as well, so why am I
discussing this as if it were a traitor change?

Well for starters, the actual applications of a crew atmos tech using
these are fairly limited.

Regular rebar bows can only fire 1 shot at a time, don't have a scope,
and you won't get much mileage out of the wall piercing component unless
you get thermals or x-rays, which are very much a stretch to acquire as
an atmos tech.

In the hands of a traitor, however, I can see it as a potential use for
a long-range sniping tool, if you are willing to commit half of your
round to make the bolts that is.

:cl:
add: A Syndicate Rebar Quiver has been added to the uplink
qol: Left clicking with a rebar crossbow will now draw/undraw the
string.
balance: Rebar quivers are now a neck slot item.
balance: Hydrogen bolts damage has been upped to 55 brute and can now
pierce through walls, they no longer have infinite piercing and can no
longer embed however.
code: removed the TRAIT_ALWAYS_HIT_ZONE, replaced with 2 new variables.
fix: fixes rebar crossbows having a higher capacity than intended if a
bolt had already been chambered.
/:cl:

---------

Co-authored-by: Xander3359 <66163761+Xander3359@users.noreply.github.com>