[NO GBP] Re-adds the cytopro to the science areas (#86441)
The Cytopro makes a grand return post-revert! Also renames the starting
petri dishes to "basic sample petri dish"

This noticeably does NOT fix the fact that the machine sells the wrong
sort of labcoat, so I strongly recommend merging #86122 immediately
following this PR.

Puts #85338 and #85996 back

:cl:
add: The vendor of cytology equipment, the CytoPro, is once again
available in your local science department!
/:cl: